test: update E2E test for .env file
